qa: fix report.html never written (un-awaited dynamic import raced process.exit)
Extract htmlReport into report-html.mjs — also a standalone regenerator:
  node skills/qa/scripts/report-html.mjs --report stardust/qa/report.json
